Remove a seemingly harmless "display: none" in CSS.
Incredibly, this CSS managed to break Drupal's tabledrag on the filter rearrange form; tabledrag.js started interpreting the table row with the
operator dropdown in each filter group (the one that is aligned left and spans all the other rows) as having a row height equal to its actual
height (around 90px), rather than 0.5px as it was interpreted before. This prevented you from dragging to any of the rows in the top half of the
filter group.

Since the "display: none" was only there to fix a 1px gap, we'll remove it and live with the 1px gap for now :)