Fix lots of bugs with the AJAX and non-JS fallbacks on the add view wizard.
Previously they did not always work correctly (and were not always consistent with each other), depending on the combination of choices that were
made on the form. This commit should fix those issues, and also adds two helper functions that make it easy to correctly add new AJAX-enabled
elements to this dynamic form.