The form piece returned by AJAX is not being run through form_builder, so #array_parents is not being set.
When #array_parents is not set to at least an empty array(), theme_container does not copy the value from ['#id'] to ['#attributes']['id'].

In this case, the styles hung off of a few ids were no longer applying because the ids were not being printed in the structure.

I fixed this by moving the ids to ['#attributes']['id'] from ['#id'] in three cases.