Revised fix for issue #2723
This adds a new field 'Block layout description' to the views UI for block displays and provides this as the description of a custom views block on the manage blocks page in Layouts. This enables each block display of a view to provide its own description together with the existing 'Block admin label' as its title.