Issue 5603. Fix views_handler_field_bulk_form
Fixes https://github.com/backdrop/backdrop-issues/issues/5603

This is an alternative PR that takes a lighter approach. Instead of changing the class parent to `views_handler_field_entity` and making other modifications to adjust to this parent, it keeps the original parent class `views_handler_field` and modifies the method `get_entity_type` to provide the correct actions for the entity type to which this bulk form field belongs.