Swap out custom code with image_style_url function Issue #6860
There is a bunch of custom code in the render function of
views_handler_field_file_uri.inc that attempts to generate the image
style url. This code breaks when you're using modules such as s3fs which
doesn't use the site default file path. I'm swapping that code out
instead with image_style_url which fixes the problem.