Issue #6877: Use label callback for labels
This came out of a discussion around the removal of the 'label callback'
setting in hook_entity_info() and some problems I was having porting
entity_translation to Backdrop. During the developer meeting, it was
decided that we would switch back to using the 'label callback' and use
it if it's set in the core entity class. This patch removes the label
methods from core classes and uses the one in the core entity class.
Doing this also necesitiated adding the 'label' property back to the
$entity_info['entity keys'] array.

Fixes #6877