Issue #4984 Prevent duplicate locale prefixes
The link module calls url() twice on the the same URL. If the site has
local enabled and is setup to detect the language with URL prefixes,
then it will add the prefix a second time to a path that already has the
language prefix. For example, it will rewrite fr/node/16 to
fr/fr/node/16. Having the locale module not add the prefix again if it's
already there, fixes this.