Fixup tests
I had to change a few things to get the test to run -- namely after the
block hooks run it returns a string instead of an array. In order to get
the layout contextual links to be added I had to switch it back to an
array. I'm not super happy about how that code is, but I'm going to
tackle that in another commit.

Fixes https://github.com/backdrop/backdrop-issues/issues/6893