Issue #6952: Fix Locale URL rewriting breaking public file URLs
This is happening because public file URLs are not handled the same as
other routes in Backdrop and are instead handled directly by the web
server and don't pass through any of the multilingual URL handling that other
routes do. For that reason, we don't want them to get prefixed by the
langcode if the locale module is enabled and the Prefix URL
detection/selection method is in use as the prefixed URL will return a
404.

This change checks to see if the path passed in is a public file URL and
opts out of prefixing the path in that case.