Add a langcode() method to entities
langcode() is a new method that basically should replace the
entity_language method that existed in Drupal 7. This was largely
replaced with just the use of the langcode property in Backdrop but this
creates problems with entity_translation because that provides
field-based translation rather than whole node translation. We need a
way to set a handler to return the langcode rather than just grabbing
that the node row in the database.

Fixes 185